// lib/types.ts
export type FeeTier = {
  from: number;
  to: number;
  fee: number;
};
// ...
export type TransactionLimits = {
    internal: {
        unverified: { min: number, max: number },
        verified: { min: number, max: number },
        merchant: { min: number, max: number },
    },
    egypt: {
        instapay: { min: number, max: number },
        wallet: { min: number, max: number },
        delivery: { min: number, max: number },
    }
};

export type AppSettings = {
    banners: string[];
    supportNumbers: {
        libyan: string;
        egyptian: string;
    };
    regions: { [key: string]: Region };
};

export type FeeSettings = {
    internal: { [key: string]: FeeTier };
    wallet: { [key: string]: FeeTier };
    instapay: { [key: string]: FeeTier };
    delivery: { [key: string]: FeeTier };
    internalFreeTransactions: number;
};